* 'ready' of https://github.com/WeNeedName/WeNeedName_backend:
  test코드 작성 및 Images entity 비활성화
  line 그래프 수정
  통계 날자 포멧 수정

+++ src/main/java/com/example/formproject/dto/response/LineChartDataDto.java
@@ -19,9 +19,9 @@ public class LineChartDataDto {
 
     @Builder.Default
     @Schema(type = "List",example = "[100,200]")
-    private List<Integer> data = new ArrayList<>();
+    private List<Long> data = new ArrayList<>();
 
-    public void addData(int data){
+    public void addData(long data){
         this.data.add(data);
     }
 }
